config.h -> use drivers section, Managers renamed to Services
