+++ OWLOS/src/main.cpp
@@ -40,7 +40,6 @@ OWLOS распространяется в надежде, что она буде
 --------------------------------------------------------------------------------------*/
 #ifndef noPlatformIO
 
-#include <Arduino.h>
 #include "src/Kernel.h"
 
 void setup()
