+++ OWLOS/src/src/drivers/DHTDriver.cpp
@@ -38,7 +38,7 @@ OWLOS распространяется в надежде, что она буде
 Вы должны были получить копию Стандартной общественной лицензии GNU вместе с
 этой программой. Если это не так, см. <https://www.gnu.org/licenses/>.)
 --------------------------------------------------------------------------------------*/
-
+#ifdef USE_DHT_DRIVER
 #include "DHTDriver.h"
 
 /*-------------------------------------------------------------------------------------------------------------------------
@@ -650,3 +650,4 @@ bool DHTDriver::writeTemperatureHistoryFile(float _historydata)
 	}
 	return result;
 }
+#endif
