+++ OWLOS/src/src/drivers/LCDDriver.h
@@ -38,11 +38,12 @@ OWLOS распространяется в надежде, что она буде
 Вы должны были получить копию Стандартной общественной лицензии GNU вместе с
 этой программой. Если это не так, см. <https://www.gnu.org/licenses/>.)
 --------------------------------------------------------------------------------------*/
+#ifdef USE_LCD_DRIVER
 
 #ifndef LCDDRIVER_H
 #define LCDDRIVER_H
 
-#include <Arduino.h>
+
 #include "../libraries\LiquidCrystal_I2C\LiquidCrystal_I2C.h" //https://www.dfrobot.com/wiki/index.php/I2C/TWI_LCD1602_Module_(Gadgeteer_Compatible)_(SKU:_DFR0063)   http://www.dfrobot.com/wiki/index.php?title=I2C/TWI_LCD1602_Module_(SKU:_DFR0063) Download: http://www.dfrobot.com/image/data/DFR0154/LiquidCrystal_I2Cv1-1.rar
 #include "BaseDriver.h"
 
@@ -128,4 +129,5 @@ private:
 	int y = 0;
 	String text = "";
 };
+#endif
 #endif
