+++ OWLOS/src/src/drivers/SensorDriver.h
@@ -39,7 +39,10 @@ OWLOS распространяется в надежде, что она буде
 этой программой. Если это не так, см. <https://www.gnu.org/licenses/>.)
 --------------------------------------------------------------------------------------*/
 
-#include <Arduino.h>
+#ifdef USE_SENSOR_DRIVER
+#ifndef SENSOR_H
+#define SENSOR_H
+
 #include "BaseDriver.h"
 
 class SensorDriver : public BaseDriver
@@ -83,3 +86,4 @@ private:
 	int data = -1;
 	float sensorTriger = 0;
 };
+#endif
