+++ OWLOS/src/src/drivers/ValveDriver.h
@@ -39,7 +39,10 @@ OWLOS распространяется в надежде, что она буде
 этой программой. Если это не так, см. <https://www.gnu.org/licenses/>.)
 --------------------------------------------------------------------------------------*/
 
-#include <Arduino.h>
+#ifdef USE_VALVE_DRIVER
+#ifndef VALVEDRIVER_H
+#define VALVEDRIVER_H
+
 #include "BaseDriver.h"
 
 #define CLOSE_PIN_INDEX 0
@@ -94,3 +97,5 @@ private:
 	int newphysicalposition = 0;
 
 };
+#endif
+#endif
