+++ OWLOS/src/src/services/PinService.cpp
@@ -40,8 +40,8 @@ OWLOS распространяется в надежде, что она буде
 --------------------------------------------------------------------------------------*/
 
 /*
-PinManager - реализует программный интерфейс между функциями доступа к физическим контактным разъемам (Pins) предоставляемыми "Arduino" библиотеками и драйверами OWLOS.
-Задачи PinManager:
+PinService - реализует программный интерфейс между функциями доступа к физическим контактным разъемам (Pins) предоставляемыми "Arduino" библиотеками и драйверами OWLOS.
+Задачи PinService:
 - управление назначением, режимами работы и доступом к Pin со стороны драйверов.
 - предоставление данных о количестве доступных Pin на микроконтроллере на котором запущена OWLOS, а так же данных о типах и режимах работы Pin.
 
@@ -55,7 +55,7 @@ PinManager - реализует программный интерфейс меж
 //---------------------------------------------------------------------------------------------------------------------------------------
 
 Указатель Pin * pins содержит ссылку на динамический массив с данными о пинах. Так как на разных типах микроконтроллеров разное количество пинов,
-массив с данными создается динамически, при вызове DriverManager.Init() - до загрузки драйверов.
+массив с данными создается динамически, при вызове DriverService.Init() - до загрузки драйверов.
 Данные пина:
 	String name = ""             - название пина, обычно используется то что напечатано на плате микроконтроллера
 	int mode = -1;               - режим работы пина Output, Input, Input-Pull Up
@@ -84,7 +84,7 @@ PinManager - реализует программный интерфейс меж
 */
 
 #include <core_version.h>
-#include "PinManager.h"
+#include "PinService.h"
 
 #ifdef ARDUINO_ESP32_RELEASE_1_0_4
 //Сторонняя библиотека реализующая analogWrite() функцию для ESP32
@@ -790,7 +790,7 @@ int parseI2CAddr(String addrStr)
 
 //декодер имен пинов в численное программное представление пина для микроконтроллеров на основе ESP8266
 //дело в том что "Arduino" ESP8266 SDK (Framework) предоставляет набор из Dx констант хранящих программные 
-//номера пинов (GPIONumber в терминологии PinManager). Для различных контроллеров эти номера отличаются. 
+//номера пинов (GPIONumber в терминологии PinService). Для различных контроллеров эти номера отличаются. 
 //Смотрите: packages\esp8266\hardware\esp8266\2.5.0\variants\ файлы pins_arduino.h определяющие номера пинов
 //
 //мы не знаем на каком именно микроконтроллере будет собрана OWLOS, но нам нужен способ получить программные номера
