+++ OWLOS/src/src/services/ScriptService.cpp
@@ -119,7 +119,7 @@ Battle Hamster Script
 */
 
 #include <Arduino.h>
-#include "DriverManager.h"
+#include "DriverService.h"
 #include "../drivers/ESPDriver.h"
 
 #define SCRIPT_ID "script"
