+++ OWLOS/src/src/transports/HTTPWebClient.cpp
@@ -1,4 +1,4 @@
-/*----------------------------------------------------------------------------
+﻿/* ----------------------------------------------------------------------------
 Ready IoT Solution - OWLOS
 Copyright 2019, 2020 by:
 - Konstantin Brul (konstabrul@gmail.com)
@@ -37,10 +37,102 @@ OWLOS распространяется в надежде, что она буде
 
 Вы должны были получить копию Стандартной общественной лицензии GNU вместе с
 этой программой. Если это не так, см. <https://www.gnu.org/licenses/>.)
---------------------------------------------------------------------------------------
---> NOTE: Для совместимости и сборки из Arduino Studion - откройте этот #define
 --------------------------------------------------------------------------------------*/
+#include <core_version.h>
 
-//#define noPlatformIO
-#include "src/src/Kernel.h"
+#ifdef ARDUINO_ESP8266_RELEASE_2_5_0
+#include <ESP8266WiFi.h>
+#include <ESP8266WiFiMulti.h>
+#include <ESP8266HTTPClient.h>
+#include <ESP8266httpUpdate.h>
+#endif
 
+#ifdef ARDUINO_ESP32_RELEASE_1_0_4
+#include <SPIFFS.h>
+#include <WiFi.h>
+#include <WiFiMulti.h>
+#include <HTTPClient.h>
+#include <HTTPUpdate.h>
+#endif
+
+#include <Arduino.h>
+#include <FS.h>
+
+#include "../utils/Utils.h"
+#include "../services/FileService.h"
+
+#define webclientid "WebClient"
+HTTPClient http;
+
+bool downloadFile(String fileName, String  url) {
+	bool result = false;
+
+#ifdef DetailedDebug 
+	debugOut(webclientid, "download: " + fileName + " from: " + url);
+#endif
+
+	if (!filesBegin())
+	{
+		//An Error has occurred while mounting file system
+		return result;;
+	}
+
+	http.begin(url);
+	int httpCode = http.GET();
+	if (httpCode > 0) {
+
+		File file = SPIFFS.open(fileName, "w");
+		if (!file) {
+#ifdef DetailedDebug 
+			debugOut(webclientid, "There was an error opening the file for writing: " + fileName);
+#endif
+			return result;
+		}
+
+		if (httpCode == HTTP_CODE_OK) {
+			int len = http.getSize();
+#ifdef DetailedDebug 
+			debugOut(webclientid, "download size: " + String(len));
+#endif
+			uint8_t buff[128] = { 0 };
+			WiFiClient * stream = http.getStreamPtr();
+			int dotCount = 0;
+			while (http.connected() && (len > 0 || len == -1)) {
+				size_t size = stream->available();
+				if (size) {
+					dotCount++;
+					if (dotCount > 40)
+					{
+						dotCount = 0;						
+					}
+					int c = stream->readBytes(buff, ((size > sizeof(buff)) ? sizeof(buff) : size));
+					file.write(buff, c);
+					if (len > 0) {
+						len -= c;
+					}
+				}
+				delay(10);
+			}
+			file.close();
+#ifdef DetailedDebug 
+			debugOut(webclientid, "download=OK");
+#endif
+			result = true;
+
+		}
+		else
+		{
+#ifdef DetailedDebug 
+			debugOut(webclientid, "download=fail HTTPResult=" + String(httpCode));
+#endif
+		}
+	}
+	else
+	{
+#ifdef DetailedDebug 
+		debugOut(webclientid, "download=fail");
+#endif
+	}
+	http.end();
+	return result;
+}
