+++ OWLOS/src/src/utils/Utils.cpp
@@ -40,7 +40,7 @@ OWLOS распространяется в надежде, что она буде
 --------------------------------------------------------------------------------------*/
 
 #include "Utils.h"
-#include "..\Managers\FileManager.h"
+#include "..\services\FileService.h"
 
 
 bool filesAtRecurse = false;
