+++ OWLStarter/src/starter.cpp
@@ -44,8 +44,8 @@ OWLOS распространяется в надежде, что она буде
   Arduino Studio Setup
   --------------------
   - Download and isnstall COM port driver for ESP8266 https://wiki.wemos.cc/downloads
-  - Download and install File\Preferenses -> Addition board managers -> http://arduino.esp8266.com/stable/package_esp8266com_index.json (NOTE: version 2.5.0)
-  * see: https://github.com/wemos/Arduino_D1 and use Tools\Board..Board manager - install D1 board support tools
+  - Download and install File\Preferenses -> Addition board services -> http://arduino.esp8266.com/stable/package_esp8266com_index.json (NOTE: version 2.5.0)
+  * see: https://github.com/wemos/Arduino_D1 and use Tools\Board..Board Service - install D1 board support tools
 
   This Sketh Setup
   ----------------
