+++ README.md
@@ -19,8 +19,8 @@
 # How to build:
   - isnstall Arduino Studio 1.8.9. or upper.
   - isnstall COM port driver if your ESP8266 board use CH340 chip https://wiki.wemos.cc/downloads
-  - at Arduino Studio menu File\Preferenses -> Addition board managers -> http://arduino.esp8266.com/stable/package_esp8266com_index.json (NOTE: version 2.5.0)
-  - at Arduino Studio menu Tools\Board->Board manager find and install ESP8266 commnodey (https://github.com/esp8266/Arduino)
+  - at Arduino Studio menu File\Preferenses -> Addition board services -> http://arduino.esp8266.com/stable/package_esp8266com_index.json (NOTE: version 2.5.0)
+  - at Arduino Studio menu Tools\Board->Board Service find and install ESP8266 commnodey (https://github.com/esp8266/Arduino)
   * see: https://github.com/wemos/Arduino_D1
   - build and upload OWL firmware on your board.
   * after uploading:
